Move FilePathUtil and DefaultFileHierarchySet to `:files`
