+++ subprojects/core-api/src/main/java/org/gradle/api/file/RelativePath.java
@@ -15,7 +15,7 @@
  */
 package org.gradle.api.file;
 
-import org.gradle.api.file.internal.FilePathUtil;
+import org.gradle.internal.file.FilePathUtil;
 
 import javax.annotation.Nullable;
 import java.io.File;
