+++ subprojects/files/src/main/java/org/gradle/internal/file/FilePathUtil.java
@@ -14,13 +14,15 @@
  * limitations under the License.
  */
 
-package org.gradle.api.file.internal;
-
-import org.apache.commons.lang.StringUtils;
+package org.gradle.internal.file;
 
 import java.io.File;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.StringTokenizer;
 
 public class FilePathUtil {
+    private static final String[] EMPTY_STRING_ARRAY = {};
     // On Windows, / and \ are separators, on Unix only / is a separator.
     private static final String FILE_PATH_SEPARATORS = File.separatorChar != '/' ? ("/" + File.separator) : File.separator;
 
@@ -28,7 +30,12 @@ public class FilePathUtil {
     }
 
     public static String[] getPathSegments(String path) {
-        return StringUtils.split(path, FILE_PATH_SEPARATORS);
+        StringTokenizer tokenizer = new StringTokenizer(path, FILE_PATH_SEPARATORS);
+        List<String> segments = new ArrayList<String>();
+        while (tokenizer.hasMoreElements()) {
+            segments.add(tokenizer.nextToken());
+        }
+        return segments.toArray(EMPTY_STRING_ARRAY);
     }
 
     /**
