+++ subprojects/files/src/test/groovy/org/gradle/internal/file/DefaultFileHierarchySetTest.groovy
